+++ CHANGELOG.rst
@@ -1,3 +1,13 @@
+0.1.12
+------
+
+* add: -l option for ``bothub ls``
+* add: --version option
+* add: check latest version
+* add: check auth token expiration
+* tune messages
+
+
 0.1.11
 ------